requirements.txt
requests==2.28.1
feedparser==6.0.10
beautifulsoup4==4.11.1
selenium==4.8.0
psycopg2-binary==2.9.5
fastapi==0.88.0
uvicorn==0.20.0